moved some parts to the lib folder
