+++ platformio.ini
@@ -12,7 +12,8 @@
 platform = teensy
 board = teensy41
 framework = arduino
+lib_dir = 
+    can_comms
 build_flags = 
-    -I can/
     -I exec/
     -I util/
